requirements.txt
openai==1.96.0
pillow==10.1.0
phonenumbers==8.13.25
